refactor:更改文件结构
更改核心模块为core，便于和其他组合作实现模块松耦合
