+++ core/Circle.cpp
