+++ core/Circle.h
