+++ core/IntersectProject2.cpp
@@ -16,7 +16,7 @@ bool myequal(const pair<double, double> pair1, const pair<double, double> pair2)
 	return (fabs(pair1.first - pair2.first) < eps && fabs(pair1.second - pair2.second) < eps);
 }
 
-void solve(vector<pair<double, double>> &realIntersections) throw(const char*)
+void __cdecl solve(vector<pair<double, double>> &realIntersections) throw(const char*)
 {
 	vector<pair<double, double>> intersections;
 	for (unsigned i = 0; i < lines.size(); i++)
@@ -69,7 +69,7 @@ void solve(vector<pair<double, double>> &realIntersections) throw(const char*)
 	//}
 }
 
-void ioHandler(string input)throw(const char*)
+void __cdecl ioHandler(string input)throw(const char*)
 {
 	fstream inputfile(input);
 	int n;
@@ -117,7 +117,7 @@ void ioHandler(string input)throw(const char*)
 	}
 }
 
-void addLine(double x1, double y1, double x2, double y2, int type)throw(const char*)
+void __cdecl addLine(double x1, double y1, double x2, double y2, int type)throw(const char*)
 {
 	if (fabs(x1) > 100000 || fabs(y1) > 100000 || fabs(x2) > 100000 || fabs(y2) > 100000)
 	{
@@ -130,7 +130,7 @@ void addLine(double x1, double y1, double x2, double y2, int type)throw(const ch
 	lines.push_back(Line(x1, y1, x2, y2, type));
 }
 
-void deleteLine(double x1, double y1, double x2, double y2, int type)
+void __cdecl deleteLine(double x1, double y1, double x2, double y2, int type)
 {
 	for (auto iter = lines.begin(); iter!= lines.end();iter++)
 	{
@@ -142,7 +142,7 @@ void deleteLine(double x1, double y1, double x2, double y2, int type)
 	}
 }
 
-void addCircle(double c1, double c2, double r)throw(const char*)
+void __cdecl addCircle(double c1, double c2, double r)throw(const char*)
 {
 	if (fabs(c1) > 100000 || fabs(c2) > 100000)
 	{
@@ -155,7 +155,7 @@ void addCircle(double c1, double c2, double r)throw(const char*)
 	circles.push_back(Circle(c1, c2, r));
 }
 
-void deleteCircle(double c1, double c2, double r)
+void __cdecl deleteCircle(double c1, double c2, double r)
 {
 	for (auto iter = circles.begin(); iter != circles.end(); iter++)
 	{
@@ -167,7 +167,7 @@ void deleteCircle(double c1, double c2, double r)
 	}
 }
 
-void deleteAll()
+void __cdecl deleteAll()
 {
 	lines.clear();
 	circles.clear();
