+++ core/const.h
